Fixes some of the codeclimate issues
