+++ services/api-gateway/src/simcore_service_api_gateway/auth_security.py
@@ -3,9 +3,7 @@
 """
 import logging
 import os
-import subprocess
 from datetime import datetime, timedelta
-from subprocess import CalledProcessError, CompletedProcess
 from typing import Dict, List, Optional
 
 import jwt
@@ -19,15 +17,16 @@ from .schemas import TokenData, UserInDB
 log = logging.getLogger(__name__)
 
 # PASSWORDS
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+__pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
 
 
 def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
+    return __pwd_context.verify(plain_password, hashed_password)
 
 
 def get_password_hash(password: str) -> str:
-    return pwd_context.hash(password)
+    return __pwd_context.hash(password)
 
 
 def authenticate_user(username: str, password: str) -> Optional[UserInDB]:
@@ -37,18 +36,6 @@ def authenticate_user(username: str, password: str) -> Optional[UserInDB]:
     if not verify_password(password, user.hashed_password):
         return None
     return user
-
-
-def create_secret_key() -> str:
-    # NOTICE that this key is reset when server is restarted!
-    try:
-        proc: CompletedProcess = subprocess.run(
-            "openssl rand -hex 32", check=True, shell=True
-        )
-    except (CalledProcessError, FileNotFoundError) as why:
-        raise ValueError(f"Cannot create secret key") from why
-    log.warning("Created new secret key!!")
-    return str(proc.stdout).strip()
 
 
 # JSON WEB TOKENS (JWT)
@@ -80,9 +67,8 @@ def decode_token(encoded_jwt: str) -> Dict:
 
 def get_access_token_data(encoded_jwt: str) -> Optional[TokenData]:
     """
-        Decodes and validates
+        Decodes and validates JWT and returns TokenData
     """
-    # returns valid
     try:
         # decode JWT [header.payload.signature] and get payload:
         payload: Dict = decode_token(encoded_jwt)
