+++ services/api-gateway/src/simcore_service_api_gateway/utils/helpers.py
@@ -1,8 +1,10 @@
-from typing import Dict
 import json
+import subprocess
 from datetime import datetime
+from subprocess import CalledProcessError, CompletedProcess  # nosec
+from typing import Dict
 
-# CAST
+
 def to_bool(s: str) -> bool:
     return s.lower() in ["true", "1", "yes"]
 
@@ -17,3 +19,15 @@ def jsoncoverter(obj):
 
 def json_dumps(obj: Dict) -> str:
     return json.dumps(obj, indent=2, default=jsoncoverter)
+
+
+def create_secret_key() -> str:
+    # NOTICE that this key is reset when server is restarted!
+
+    try:
+        proc: CompletedProcess = subprocess.run(            # nosec
+            "openssl rand -hex 32", check=True, shell=True
+        )
+    except (CalledProcessError, FileNotFoundError) as why:
+        raise ValueError(f"Cannot create secret key") from why
+    return str(proc.stdout).strip()
