+++ services/api-gateway/tools/gen_api.py
@@ -38,7 +38,7 @@ def as_class_name(name):
 
 # templates
 template_env = Environment(
-    autoescape=False, loader=FileSystemLoader(current_dir / "templates")
+    autoescape=False, loader=FileSystemLoader(current_dir / "templates") # nosec
 )
 template_env.globals.update({"len": len, "cls_name": as_class_name})
